Refa: move Safe Core SDK-related files to services/safe-core
